added header to AllWordsListRV

+++ app/src/main/java/com/example/lernen/ui/addNewWord/AddNewWordViewModel.kt
@@ -33,7 +33,7 @@ constructor(
         Log.d("AddNewWordViewModel", "Word - ${newWord.value}")
         Log.d("AddNewWordViewModel", "Translation - ${newTranslation.value}")
         viewModelScope.launch {
-            mainRepository.setWord(WordEntity(0, newLesson.value!!.toInt(), newWord.value!!, newTranslation.value!!))
+            mainRepository.setWord(WordEntity(0, newLesson.value!!, newWord.value!!, newTranslation.value!!))
             Log.d("AddNewWordViewModel", "new word added")
             newLesson.value = ""
             newWord.value = ""
